requirements.txt
django
djangorestframework
djangorestframework-simplejwt
PyJWT
django-qrcode
cryptography
Pillow
celery 
redis 
django-celery-beat
gevent
python-dotenv
